requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
PyPDF2==3.0.0
python-docx==1.1.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.10
email-validator==2.2.0
spacy==3.7.6
reportlab==4.3.1
